backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pymysql
pydantic
psycopg2-binary
pymongo
neo4j
pandas
mlxtend
requests
python-dotenv
email-validator
polars
pylint